requirements.txt
langchain
langchain-community
langchain-core
openai
langchain-openai
huggingface

pandas
python-dotenv
numpy

streamlit